Modified Woo screen UI into multiple groups
