+++ example/src/main/java/org/wordpress/android/fluxc/example/FragmentExt.kt
@@ -9,3 +9,10 @@ import android.support.v4.app.Fragment
 fun Fragment.prependToLog(s: String) {
     (activity as? MainExampleActivity)?.prependToLog(s)
 }
+
+fun Fragment.replaceFragment(fragment: Fragment) {
+    fragmentManager?.beginTransaction()
+            ?.replace(R.id.fragment_container, fragment)
+            ?.addToBackStack(null)
+            ?.commit()
+}
